# http-handler/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/court/all": {
            "get": {
                "description": "Retourne la liste de tous les terrains enregistrés en base de données",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "terrain"
                ],
                "summary": "Liste tous les terrains",
                "responses": {
                    "200": {
                        "description": "Liste des terrains",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DBCourt"
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur lors de la récupération des terrains",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
        "/court/{id}": {
            "get": {
                "description": "Retourne les informations d’un terrain (court) en fonction de son identifiant passé dans l’URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "terrain"
                ],
                "summary": "Récupère un terrain par son ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identifiant du terrain",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Terrain trouvé",
                        "schema": {
                            "$ref": "#/definitions/models.DBCourt"
                        }
                    },
                    "400": {
                        "description": "ID manquant ou invalide",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "404": {
                        "description": "Terrain non trouvé",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    },
                    "500": {
                        "description": "Erreur serveur ou base de données",
                        "schema": {
                            "$ref": "#/definitions/models.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DBCourt": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}